refactor: Add admin access to get_cards_by_user and get_card routes
This commit modifies the `card_controller.py` file to include admin access for the `get_cards_by_user` and `get_card` routes. If the user is an admin, these routes will return all cards instead of just the ones associated with the user. This change improves the functionality and flexibility of the application.